+++ src/core/component/init/test/b-test-component/b-test-component.ts
@@ -1,0 +1,15 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+import { component } from 'core/component';
+import iBlock from 'components/super/i-block/i-block';
+
+@component()
+export default class bTestComponent extends iBlock {
+
+}