+++ src/core/component/init/test/b-test-component/index.js
@@ -1,0 +1,11 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+package('b-test-component')
+	.extends('i-block')
+	.dependencies('b-button');