+++ CHANGELOG.md
@@ -15,7 +15,8 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 #### :bug: Bug Fix
 
-* Fix hookChange event processing for activate/deactivate component `core/component/init`
+* Fixed a typo in the hookChange event name for processing activate/deactivate component `core/component/init`
+* Fix component deactivation order, children must be deactivated first `core/component/init`
 * Fix hookChange event should not bubble `bDynamicPage`
 
 ## v4.0.0-beta.106 (2024-06-25)