+++ src/core/component/init/CHANGELOG.md
@@ -13,7 +13,8 @@ Changelog
 
 #### :bug: Bug Fix
 
-* Fix hookChange event processing for activate/deactivate component
+* Fixed a typo in the hookChange event name for processing activate/deactivate component
+* Fix component deactivation order, children must be deactivated first
 
 ## v4.0.0-beta.91 (2024-04-19)